+++ src/spnego/_ntlm.py
@@ -18,6 +18,13 @@ from spnego._context import (
     WrapResult,
     split_username,
 )
+from spnego._credential import (
+    Credential,
+    CredentialCache,
+    NTLMHash,
+    Password,
+    unify_credentials,
+)
 from spnego._ntlm_raw.crypto import (
     RC4Handle,
     compute_response_v1,
@@ -196,19 +203,26 @@ def _get_workstation() -> typing.Optional[str]:
 class _NTLMCredential:
     def __init__(
         self,
-        domain: typing.Optional[str] = None,
-        username: typing.Optional[str] = None,
-        password: typing.Optional[str] = None,
+        credential: typing.Optional[Credential] = None,
     ) -> None:
         self._store: typing.Optional[str]
-        if password:
-            self.domain = domain
-            self.username = username
-            self.lm_hash = lmowfv1(password)
-            self.nt_hash = ntowfv1(password)
+        if isinstance(credential, Password):
             self._store = "explicit"
+            self.domain, self.username = split_username(credential.username)
+            self.lm_hash = lmowfv1(credential.password)
+            self.nt_hash = ntowfv1(credential.password)
+
+        elif isinstance(credential, NTLMHash):
+            self._store = "explicit"
+            self.domain, self.username = split_username(credential.username)
+            self.lm_hash = base64.b16decode(credential.lm_hash.upper()) if credential.lm_hash else b"\x00" * 16
+            self.nt_hash = base64.b16decode(credential.nt_hash.upper()) if credential.nt_hash else b"\x00" * 16
 
         else:
+            domain = username = None
+            if isinstance(credential, CredentialCache):
+                domain, username = split_username(credential.username)
+
             self._store = _get_credential_file()
             self.domain, self.username, self.lm_hash, self.nt_hash = _get_credential(self._store, domain, username)
 
@@ -221,8 +235,8 @@ class NTLMProxy(ContextProxy):
 
     def __init__(
         self,
-        username: typing.Optional[str],
-        password: typing.Optional[str],
+        username: typing.Optional[typing.Union[str, Credential, typing.List[Credential]]] = None,
+        password: typing.Optional[str] = None,
         hostname: typing.Optional[str] = None,
         service: typing.Optional[str] = None,
         channel_bindings: typing.Optional[GssChannelBindings] = None,
@@ -230,11 +244,11 @@ class NTLMProxy(ContextProxy):
         usage: str = "initiate",
         protocol: str = "ntlm",
         options: NegotiateOptions = NegotiateOptions.none,
-        _is_wrapped: bool = False,
         **kwargs: typing.Any,
     ) -> None:
+        credentials = unify_credentials(username, password, required_protocol="ntlm")
         super(NTLMProxy, self).__init__(
-            username, password, hostname, service, channel_bindings, context_req, usage, protocol, options, _is_wrapped
+            credentials, hostname, service, channel_bindings, context_req, usage, protocol, options
         )
 
         self._complete = False
@@ -270,8 +284,7 @@ class NTLMProxy(ContextProxy):
             self._context_req &= ~NegotiateFlags.extended_session_security
 
         if self.usage == "initiate":
-            domain, user = split_username(self.username)
-            self._credential = _NTLMCredential(domain=domain, username=user, password=self.password)
+            self._credential = _NTLMCredential(next(c for c in credentials if "ntlm" in c.supported_protocols))
 
             self._lm = lm_compat_level < 2
             self._nt_v1 = lm_compat_level < 3
@@ -474,7 +487,10 @@ class NTLMProxy(ContextProxy):
         ):
             raise OperationNotAvailableError(context_msg="Anonymous user authentication not implemented")
 
-        self._credential = _NTLMCredential(domain=auth.domain_name, username=auth.user_name)
+        username = auth.user_name
+        if auth.domain_name:
+            username = f"{auth.domain_name}\\{username}"
+        self._credential = _NTLMCredential(CredentialCache(username=username))
         expected_mic = None
 
         if auth.nt_challenge_response and len(auth.nt_challenge_response) > 24:
@@ -581,11 +597,8 @@ class NTLMProxy(ContextProxy):
         return WinRMWrapResult(header=enc_data[:16], data=enc_data[16:], padding_length=0)
 
     def unwrap(self, data: bytes) -> UnwrapResult:
-        if not self._handle_in:
-            raise NoContextError(context_msg="NTLM context has not been completed")
-
         signature = data[:16]
-        msg = self._handle_in.update(data[16:])
+        msg = self._handle_in.update(data[16:])  # type: ignore[union-attr]
         self.verify(msg, signature)
 
         return UnwrapResult(data=msg, encrypted=True, qop=0)